Fix Clock Mode for Scoreboard/Shotclock
When Clock mode is enabled by pressing for more than 3s the controller Time On/Off button, the clock data shown in the controller is not replicated in the shotclock and main display time. Also, other displays must also be switched off at the start (including the controller) unless any other buttons is pressed.